feat: agrego total inveertido en vista de plazos fijos
